rename: 매거진 컴포넌트 containers로 이전
Related to: #32
