+++ src/containers/magazine/CurrentHighlights.tsx
