+++ src/containers/magazine/MReportCoustomAlert.tsx
