+++ src/containers/magazine/MagazineHeader.tsx
