+++ src/containers/magazine/MagazineList.tsx
